transaction with kafka and facturation

+++ Backend/src/main/java/com/example/backend/repositories/FacturesRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repositories;
 import com.example.backend.model.Factures;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturesRepository extends JpaRepository<Factures, Integer> {
+public interface FacturesRepository extends JpaRepository<Factures, Long> {
 }
